requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
boto3==1.18.25
botocore==1.21.25
bs4==0.0.1
Django==3.2.5
html2image==2.0.1
jmespath==0.10.0
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
